requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==3.2
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-environ==0.9.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
stripe==5.4.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
gunicorn 
google-api-core
google-auth
google-auth-oauthlib
google-cloud
google-cloud-secret-manager
google-cloud-storage
